fix: AES_GCM implementation
