+++ KeychainExample/e2e/testCases/acessControlTest.spec.js
@@ -114,7 +114,7 @@ describe('Access Control', () => {
     );
 
     it(
-      'should save and retrieve username and password without biometrics' +
+      'should save and retrieve username and password without biometrics - ' +
         type,
       async () => {
         await expect(element(by.text('Keychain Example'))).toExist();
@@ -136,12 +136,18 @@ describe('Access Control', () => {
       }
     );
 
-    it('should retrieve username and password after app launch without biometrics', async () => {
-      await expect(element(by.text('Keychain Example'))).toExist();
-      await expect(element(by.text('hasGenericPassword: true'))).toBeVisible();
-      await element(by.text('Load')).tap();
-      await matchLoadInfo('testUsernameAny', 'testPasswordAny');
-    });
+    it(
+      'should retrieve username and password after app launch without biometrics - ' +
+        type,
+      async () => {
+        await expect(element(by.text('Keychain Example'))).toExist();
+        await expect(
+          element(by.text('hasGenericPassword: true'))
+        ).toBeVisible();
+        await element(by.text('Load')).tap();
+        await matchLoadInfo('testUsernameAny', 'testPasswordAny');
+      }
+    );
   });
 
   it('should reset all credentials', async () => {
