+++ android/src/main/java/com/oblador/keychain/cipherStorage/CipherStorageKeystoreAesGcm.kt
@@ -14,14 +14,18 @@ import com.oblador.keychain.decryptionHandler.DecryptionResultHandler
 import com.oblador.keychain.exceptions.CryptoFailedException
 import com.oblador.keychain.exceptions.KeyStoreAccessException
 import java.io.IOException
+import java.io.InputStream
 import java.security.GeneralSecurityException
 import java.security.Key
+import java.security.SecureRandom
 import java.security.spec.KeySpec
 import java.util.concurrent.atomic.AtomicInteger
 import javax.crypto.Cipher
 import javax.crypto.KeyGenerator
 import javax.crypto.SecretKey
 import javax.crypto.SecretKeyFactory
+import javax.crypto.spec.IvParameterSpec
+import javax.crypto.spec.GCMParameterSpec
 
 
 @TargetApi(Build.VERSION_CODES.M)
@@ -226,6 +230,55 @@ class CipherStorageKeystoreAesGcm(@NonNull reactContext: ReactApplicationContext
   // endregion
 
   // region Initialization Vector encrypt/decrypt support
+
+  /** Initialization vector support. */
+  object IV {
+    /** Encryption/Decryption initialization vector length. */
+    const val IV_LENGTH = 12
+    const val TAG_LENGTH = 128
+
+    /** Save Initialization vector to output stream. */
+    val encrypt = EncryptStringHandler { cipher, key, output ->
+      val iv = ByteArray(IV_LENGTH)
+
+      SecureRandom().nextBytes(iv)
+      val gcmParamSpec = GCMParameterSpec(TAG_LENGTH, iv)
+
+      cipher.init(Cipher.ENCRYPT_MODE, key, gcmParamSpec)
+      output.write(iv, 0, iv.size)
+    }
+
+    /** Read initialization vector from input stream and configure cipher by it. */
+    val decrypt = DecryptBytesHandler { cipher, key, input ->
+      val iv = readIv(input)
+      cipher.init(Cipher.DECRYPT_MODE, key, iv)
+    }
+
+    /** Extract initialization vector from provided bytes array. */
+    @Throws(IOException::class)
+    fun readIv(bytes: ByteArray): GCMParameterSpec {
+      val iv = ByteArray(IV_LENGTH)
+
+      if (IV_LENGTH >= bytes.size)
+        throw IOException("Insufficient length of input data for IV extracting.")
+
+      System.arraycopy(bytes, 0, iv, 0, IV_LENGTH)
+
+      return GCMParameterSpec(TAG_LENGTH, iv)
+    }
+
+    /** Extract initialization vector from provided input stream. */
+    @Throws(IOException::class)
+    fun readIv(inputStream: InputStream): GCMParameterSpec {
+      val iv = ByteArray(IV_LENGTH)
+      val result = inputStream.read(iv, 0, IV_LENGTH)
+
+      if (result != IV_LENGTH) throw IOException("Input stream has insufficient data.")
+
+      return GCMParameterSpec(TAG_LENGTH, iv)
+    }
+  }
+
   @NonNull
   @Throws(GeneralSecurityException::class, IOException::class)
   override fun encryptString(@NonNull key: Key, @NonNull value: String): ByteArray =
@@ -235,5 +288,6 @@ class CipherStorageKeystoreAesGcm(@NonNull reactContext: ReactApplicationContext
   @Throws(GeneralSecurityException::class, IOException::class)
   override fun decryptBytes(@NonNull key: Key, @NonNull bytes: ByteArray): String =
     decryptBytes(key, bytes, IV.decrypt)
+
   // endregion
 }
